Including GloboNetwork plugin

+++ plugins/network-elements/globonetwork/src/com/globo/globonetwork/cloudstack/guru/GloboNetworkGuru.java
@@ -16,6 +16,8 @@
 */
 package com.globo.globonetwork.cloudstack.guru;
 
+import java.util.List;
+
 import javax.ejb.Local;
 import javax.inject.Inject;
 import javax.naming.ConfigurationException;
@@ -26,9 +28,11 @@ import org.springframework.stereotype.Component;
 import com.cloud.dc.DataCenter.NetworkType;
 import com.cloud.deploy.DeployDestination;
 import com.cloud.deploy.DeploymentPlan;
+import com.cloud.exception.ConcurrentOperationException;
 import com.cloud.exception.InsufficientAddressCapacityException;
 import com.cloud.exception.InsufficientNetworkCapacityException;
 import com.cloud.exception.InsufficientVirtualNetworkCapcityException;
+import com.cloud.exception.ResourceUnavailableException;
 import com.cloud.network.Network;
 import com.cloud.network.Network.Provider;
 import com.cloud.network.NetworkProfile;
@@ -37,17 +41,22 @@ import com.cloud.network.PhysicalNetwork.IsolationMethod;
 import com.cloud.network.dao.NetworkVO;
 import com.cloud.network.guru.GuestNetworkGuru;
 import com.cloud.network.guru.NetworkGuru;
+import com.cloud.network.router.VirtualRouter;
 import com.cloud.network.router.VpcVirtualNetworkApplianceManager;
 import com.cloud.offering.NetworkOffering;
 import com.cloud.user.Account;
 import com.cloud.user.AccountManager;
+import com.cloud.user.User;
 import com.cloud.utils.db.Transaction;
 import com.cloud.utils.db.TransactionCallbackWithException;
 import com.cloud.utils.db.TransactionStatus;
 import com.cloud.utils.exception.CloudRuntimeException;
 import com.cloud.vm.NicProfile;
+import com.cloud.vm.NicVO;
 import com.cloud.vm.ReservationContext;
 import com.cloud.vm.VirtualMachineProfile;
+import com.globo.globonetwork.cloudstack.GloboNetworkVipAccVO;
+import com.globo.globonetwork.cloudstack.dao.GloboNetworkVipAccDao;
 import com.globo.globonetwork.cloudstack.manager.GloboNetworkService;
 
 @Component
@@ -63,6 +72,8 @@ public class GloboNetworkGuru extends GuestNetworkGuru {
     VpcVirtualNetworkApplianceManager _routerMgr;
     @Inject
     AccountManager _accountMgr;
+    @Inject
+    GloboNetworkVipAccDao _globoNetworkVipDao;
 
     protected NetworkType _networkType = NetworkType.Advanced;
 
@@ -167,6 +178,12 @@ public class GloboNetworkGuru extends GuestNetworkGuru {
 
         s_logger.debug("Asking GuestNetworkGuru to deallocate NIC " + nic.toString() + " from VM " + vm.getInstanceName());
 
+        long networkId = nic.getNetworkId();
+        List<GloboNetworkVipAccVO> vips = _globoNetworkVipDao.findByNetwork(networkId);
+        for (GloboNetworkVipAccVO vip: vips) {
+            NicVO nicVO = _nicDao.findById(nic.getId());
+            _globoNetworkService.disassociateNicFromVip(vip.getGloboNetworkVipId(), nicVO);
+        }
         _globoNetworkService.unregisterNicInGloboNetwork(nic, vm);
 
         super.deallocate(network, nic, vm);
@@ -174,13 +191,29 @@ public class GloboNetworkGuru extends GuestNetworkGuru {
 
     @Override
     public void shutdown(NetworkProfile profile, NetworkOffering offering) {
-        s_logger.debug("Removing networks from GloboNetwork");
-        _globoNetworkService.removeNetworkFromGloboNetwork(profile);
+        List<GloboNetworkVipAccVO> vips = _globoNetworkVipDao.findByNetwork(profile.getId());
+        if (vips != null && !vips.isEmpty()) {
+            throw new CloudRuntimeException("There is VIPs related to this network. Network destroyed will be aborted. Delete VIP before.");
+        }
 
-        s_logger.debug("Asking GuestNetworkGuru to shutdown network " + profile.getName());
-        // never call super.shutdown because it clear broadcastUri, and sometimes this
-        // make same networks without vlan
-        //super.shutdown(profile, offering);
+        try {
+            List<VirtualRouter> routers = _routerMgr.getRoutersForNetwork(profile.getId());
+            for (VirtualRouter router: routers) {
+                _routerMgr.destroyRouter(router.getId(), _accountMgr.getAccount(Account.ACCOUNT_ID_SYSTEM), User.UID_SYSTEM);
+            }
+
+            s_logger.debug("Removing networks from GloboNetwork");
+            _globoNetworkService.removeNetworkFromGloboNetwork(profile);
+
+            s_logger.debug("Asking GuestNetworkGuru to shutdown network " + profile.getName());
+            // never call super.shutdown because it clear broadcastUri, and sometimes this
+            // make same networks without vlan
+            //super.shutdown(profile, offering);
+        } catch (ResourceUnavailableException e) {
+            throw new CloudRuntimeException(e);
+        } catch (ConcurrentOperationException e) {
+            throw new CloudRuntimeException(e);
+        }
     }
 
     @Override
